+++ packages/grafana-data/src/types/annotations.ts
@@ -1,14 +1,16 @@
 import { Observable } from 'rxjs';
 import { ComponentType } from 'react';
 
-import { DataQuery, QueryEditorProps } from './datasource';
+import { QueryEditorProps } from './datasource';
 import { DataFrame } from './dataFrame';
+import { DataQuery, DatasourceRef } from './query';
 
 /**
  * This JSON object is stored in the dashboard json model.
  */
 export interface AnnotationQuery<TQuery extends DataQuery = DataQuery> {
-  datasource?: string | null;
+  datasource?: DatasourceRef | string | null;
+
   enable: boolean;
   name: string;
   iconColor: string;
